+++ CHANGELOG.md
@@ -1,3 +1,38 @@
+## [v0.4.7](https://github.com/higlass/higlass-python/compare/v0.4.7...v0.4.6)
+
+- Bumped pillow version
+
+## [v0.4.6](https://github.com/higlass/higlass-python/compare/v0.4.6...v0.4.5)
+
+- Added inline_tile support for bed-like items
+- Bumped React version dependency to ensure compatibility with latest  Juptyerlab
+- Bumped higlass version to v1.11.8
+- Bumped matplotlib version and added pillow dependency version
+
+## [v0.4.5](https://github.com/higlass/higlass-python/compare/v0.4.5...v0.4.4)
+
+- Add constructor for multivec tilesets
+- Added wrapper for bam tileset generator
+- Updated docs to include examples for bam and multivec tilesets
+- Add proxy support
+- Add support for listening on a unix socket
+
+## [v0.4.4](https://github.com/higlass/higlass-python/compare/v0.4.4...v0.4.3)
+
+-   Added default positions for non `horizontal-` tracks
+-   Update docs to include Views import in simple example
+-   Bump HiGlass to v1.11
+
+## [v0.4.2](https://github.com/higlass/higlass-python/compare/v0.4.2...v0.4.1)
+
+-   Only load FUSE python package when it is called to avoid libfuse missing error
+
+## [v0.4.1](https://github.com/higlass/higlass-python/compare/v0.4.1...v0.4.0)
+
+-   Add missing `pixi.js` dependency from `higlass-jupyter`
+-   Update `@jupyter-widgets/base` to support Jupyter Lab v2
+-   Update simple-httpfs to fix the missing `boto3` error
+
 ## [v0.4.0](https://github.com/higlass/higlass-python/compare/v0.4.0...v0.3.0)
 
 -   API for adding genome position search box within python